--- a/calculate_average_yasyf.sh
+++ b/calculate_average_yasyf.sh
@@ -15,5 +15,8 @@
 #  limitations under the License.
 #
 
-JAVA_OPTS="--enable-preview -da -dsa -XX:+UnlockExperimentalVMOptions -XX:+UseEpsilonGC -Xms512m -Xmx512m -XX:+AlwaysPreTouch"
+source "$HOME/.sdkman/bin/sdkman-init.sh"
+sdk use java 21.0.1-graal 1>&2
+
+JAVA_OPTS="--enable-preview -da -dsa -XX:+UnlockExperimentalVMOptions -XX:+UseEpsilonGC -Xms512m -Xmx512m -XX:+AlwaysPreTouch -XX:+UseTransparentHugePages"
 time java $JAVA_OPTS --class-path target/average-1.0.0-SNAPSHOT.jar dev.morling.onebrc.CalculateAverage_yasyf