```shell
$ hyperfine --warmup 1 --runs 3 "./calculate_average_yasyf.sh"
Benchmark 1: ./calculate_average_yasyf.sh
  Time (mean ± σ):      1.583 s ±  0.025 s    [User: 19.890 s, System: 1.112 s]
  Range (min … max):    1.569 s …  1.612 s    3 runs
```

@merykitty's submission:

```shell
$ hyperfine --warmup 1 --runs 3 "./calculate_average_merykitty.sh"
Benchmark 1: ./calculate_average_merykitty.sh
  Time (mean ± σ):      3.620 s ±  0.199 s    [User: 44.306 s, System: 1.311 s]
  Range (min … max):    3.498 s …  3.849 s    3 runs
```

- [x] Tests pass (`./test.sh <username>`)
- [x] All formatting changes by the build are committed
- [x] Output matches that of calculate_average_baseline.sh
* Execution time: 1.583s
* Execution time of reference implementation: 2m42.5s