@sharpobject 2024-01-07
sum can overflow on legal inputs. also, looks like this is ~100% faster than that other submission.