@yasyf 2024-01-07
@gunnarmorling ran it on a test-like box, results are less dramatic than local,  but still not bad:

```shell
[build@control 1brc]$ hyperfine --warmup 1 --runs 3 "./calculate_average_yasyf.sh"
Benchmark 1: ./calculate_average_yasyf.sh
  Time (mean ± σ):      6.493 s ±  0.126 s    [User: 43.337 s, System: 1.464 s]
  Range (min … max):    6.348 s …  6.580 s    3 runs
```