@yasyf 2024-01-09
@gunnarmorling [vectorizing](https://github.com/gunnarmorling/1brc/pull/204/commits/f9d4446234951080ba91a3fa4ef0b4623197f972) helped... a lot... can you confirm these results? running this on a similar instance.

```shell
[build@control 1brc]$ hyperfine --warmup 1 --runs 3 "./calculate_average_yasyf.sh"
Benchmark 1: ./calculate_average_yasyf.sh
  Time (mean ± σ):      1.280 s ±  0.017 s    [User: 5.856 s, System: 0.256 s]
  Range (min … max):    1.261 s …  1.294 s    3 runs
```