@yasyf 2024-01-15
@gunnarmorling sorry been a crazy week! will come back to this one tomorrow :)